+++ CHANGELOG.md
@@ -16,6 +16,7 @@ will ensure that all necessary changes have been applied before updating to 4.x.
 
 - [Add support for attributes in farmOS plugin types #963](https://github.com/farmOS/farmOS/pull/963)
 - [Add a hook for excluding fields from CSV importers #958](https://github.com/farmOS/farmOS/pull/958)
+- [Include config fields in CSV importers #959](https://github.com/farmOS/farmOS/pull/959)
 
 ### Changed